Cleanup html5 in templates
The html5 in the redlounge templates was a bit off. Mimic'ed what's in the spf13 site for wrapping the main content on any page in a section with an id of main (though this might be able to be re-done as just a <main> tag), and putting "article" around each main item in the list. If the article also has content like a summary and other metadata, separated it out into header and footer tags.