Topic list in listtop partial
Worked on the topic list in listtop.html partial. Tried to use delimit to avoid having the trailing comma, but, it does not work. Switched to using a heavy asterisk before and after, so it at least looks ok. Used css class .decorative-marker2 to color the asterisk grey, and adjust pre- and post- spacing.

Tried a couple things to show various variables in a template, but it largely does not work. Probably the context was wrong.

Worked on footer menu. All caps, wider spacing of text.