Add highlight.js, more listtops, html comments
Added highlight.js, though it was already there. Wanted to have it stored locally in my static. Set ``highlightStyle = "solarized_dark"``, because that one seems to go with the theme.

Added a couple code samples to a test post. You just put ``~~~somelang`` before the code block and ~~~ after. Hugo picks up the somelang and formats the code tag correctly.

Did a bunch of template work, adding a ``{{ "<!-- ENTERING blahblahtemplate -->" | safeHtml }}`` to the beginning of every template and partial, and, a LEAVING one to the end. This way I can see what is going on by viewing source.

Tried to set background color to html, body etc, but nothing would get the exact square where the content it. So I put a greyish background on the upper photo banner on the top page.