Worked home page blurbs, frontmatter
Home page blurbs are written. Added a photo there too.

Changed content metadata to yaml from toml, because Github renders it nicely. I like toml better, because to me it is more readable, but oh well.

You can keep the config.toml file in toml, and just reference it inside itself, and, still use yaml metadata for all the posts. This is done with the variable MetaDataFormat.