Menu update, BaseUrl to BaseURL
Found out that I can link a page to a menu, so I tried it with About, and it works. If you're using yaml frontmatter, make sure to not use a - in the menu. For whatever reason this causes chokage. This works

~~~
---
...other frontmatter schtuff...
menu:
  footer:
    Name: "Bio"
    Weight: -110
    Identifier: "about"
    Url: "/about"
---
~~~

Read that menus should be last in params in config.toml, so I moved some other params up to make that happen.

Updated BaseUrl to BaseURL to avoid whining from hugo server command. However, themes are full of these references.