Fix button link hover color, change sep symbol
Pure button link color was set to "inherit" but that causes trouble when the background is set to certain colors. I wanted to keep the colors, so I went with the theme color of magenta (ish).

Also, changed the tag separator from // per the redlounge theme, to an alembic symbol, for that which distills, and, the sidebar menu separator to a comet symbol. Just because.