Add google custom search, style blockquote
Google custom search is easy enough I suppose, but it's really not so easy to style the search box, beyond the basic things they give you. CSS styling is hit and miss for me.

If you put the js code from google in a div, then style the div, you can at least set the font of it.

There's a clear button that clears the results, but, it's not really relevant. Set it to display:none;.

Added a topic called site, for pages that have to do w/ site. Wanted to use in search.md, but, that's not in use so I deleted it.