Add logic to seo-schema partial
Added an if-else logic structure to seo-schema.html partial, so that JSON-LD is marked up differently for all posts, as well as pages like about, colophon, license.

Edited about, license, colophon frontmatter to make it work w/ the new seo-schema.html.

Added svg and png logos for creative commons, and linked them on the license page.