SEO twittercard and googlenews
Mainly more SEO partials, twittercard and googlenews.

Googlenews is probably not useful for this site.

About edited a little, to remove p tags around text after pull quote. Just put a space instead.

Added robots.txt but it's only preventing a dev scratch folder.

Moved stuff that should be in the root, to /static/, such as the google and pinterest validation htmls.