Working npm and gulp files
Started from "barebones" repo and added. This is using node,
and hence has a large number of js libraries added in
node_modules in src/. You use "npm run task" to run gulp,
which in turn makes a compressed "bundle" of css or js.

The test file is src/index.html and this looks at the bundle,
to allow some basic dev to be done.

Then, hugo looks at the bundle in static.