Fix conditional loading, clean up frontmatter, add banners
Switched logic of section loading in single.html,
by using "if not isset" to hide the section only if hidesomething is set in the frontmatter.

Cleaned up frontmatter in about, colophon, license, to
remove the various showauthor, showdate etc, and either
specify hidesomething, or, nothing at all.

Updated dates and added banners in the 3 files, as well.