Cache busting using {{ .Now.Unix }} and git hash
Bundle files get updated all the time via gulp, so they
are not so good to keep in git. If they are not in git, the
technique of using the git hash to cache bust does not
work either. Therefore, use date for those. Because it is
being done at deploy, it's a reasonable way to do it.

{{with getenv "THE_HASH"}}?v={{.}}{{end}}
Set THE_HASH on deploy before the hugo command.